+++ data/canada.js
@@ -1,5 +1,5 @@
-import { getPercent, getDayOfYear } from "./index"
-import regions from "./_regions"
+import { getPercent, getDayOfYear } from './index'
+import regions from './_regions'
 
 const canadaDays = [
   {
@@ -18,10 +18,18 @@ const canadaVaccines = [
   },
 ]
 
+const canadaFull = [
+  {
+    x: regions.CAN.abbr,
+    y: getPercent({
+      numerator: regions.CAN.full,
+      denominator: regions.CAN.population,
+    }),
+  },
+]
+
 const _getDataForRegions = (regions) => {
-  const regionsMinusCanada = Object.keys(regions).filter(
-    (abbr) => abbr !== "CAN"
-  )
+  const regionsMinusCanada = Object.keys(regions).filter((abbr) => abbr !== 'CAN')
 
   return regionsMinusCanada.map((abbr) => {
     return {
@@ -36,13 +44,13 @@ const _getDataForRegions = (regions) => {
 
 const regionVaccines = _getDataForRegions(regions)
 regionVaccines.push({
-  x: "Days in 2021",
+  x: 'Days in 2021',
   y: getPercent({ numerator: getDayOfYear(), denominator: 365 }),
-  fill: "#b9cde0",
+  fill: '#b9cde0',
 })
 
 regionVaccines.sort(function (a, b) {
   return b.y - a.y
 })
 
-export { canadaDays, canadaVaccines, regionVaccines }
+export { canadaDays, canadaFull, canadaVaccines, regionVaccines }