+++ pages/index.js
@@ -16,15 +16,26 @@ import Layout from '../components/Layout'
 import VacsLabel from '../components/VacsLabel'
 
 import { colors, theme } from '../styles/_theme'
-import { canadaDays, canadaVaccines, regionVaccines } from '../data/canada'
+import { canadaDays, canadaVaccines, canadaFull, regionVaccines } from '../data/canada'
 import {
-  _getDaysLabel,
-  _getVaccinesLabel,
-  _getDaysTooltip,
+  getDaysLabel,
+  getFullLabel,
+  getVaccinesLabel,
+  getDaysTooltip,
+  getFullTooltip,
   getRegionTooltip,
 } from '../utils/charts'
 
-const LastUpdated = () => <p>Last updated: Wednesday, Jan 20 at 6:33 pm EST.</p>
+const LastUpdated = () => (
+  <p>
+    Last updated: Wednesday, Jan 20 at 6:33 pm EST.
+    <style jsx>{`
+      margin-top: -40px;
+      margin-bottom: 40px;
+      z-index: 1;
+    `}</style>
+  </p>
+)
 
 const VacsRedLine = ({ labelY = 80, ...props }) => (
   <VictoryLine
@@ -56,12 +67,12 @@ const Home = () => (
         <h1>
           <span className="visuallyHidden">Total vaccines administered in </span>Canada
         </h1>
-        <p>Number of administered vaccines as a percentage of Canada’s population.</p>
+        <p>Percentage of Canadians who have received vaccines</p>
         <div className="chart">
-          <VictoryChart height={140} width={360} theme={theme}>
+          <VictoryChart height={154} width={360} theme={theme}>
             <VictoryLegend
               colorScale={[colors.QcOrangeAccent, colors.QcBlueLight]}
-              data={[{ name: 'Vaccinations administered' }, { name: 'Days in 2021' }]}
+              data={[{ name: 'Canadians vaccinated*' }, { name: 'Days in 2021' }]}
             />
             <VictoryAxis />
             <VictoryAxis
@@ -83,16 +94,25 @@ const Home = () => (
                 name="bar-days"
                 animate={_animateBar}
                 data={canadaDays}
-                labels={_getDaysLabel}
-                labelComponent={<VacsLabel tooltipLabel={() => _getDaysTooltip()} />}
+                labels={getDaysLabel}
+                labelComponent={<VacsLabel tooltipLabel={() => getDaysTooltip()} />}
               />
               <VictoryBar
                 animate={_animateBar}
                 name="bar-vaccines"
                 data={canadaVaccines}
-                labels={_getVaccinesLabel}
+                labels={getVaccinesLabel}
                 labelComponent={
                   <VacsLabel tooltipLabel={(label) => getRegionTooltip(label.datum.x)} />
+                }
+              />
+              <VictoryBar
+                animate={_animateBar}
+                name="bar-two"
+                data={canadaFull}
+                labels={getFullLabel}
+                labelComponent={
+                  <VacsLabel tooltipLabel={(label) => getFullTooltip(label.datum.x)} />
                 }
               />
             </VictoryGroup>
@@ -102,14 +122,22 @@ const Home = () => (
 
         <LastUpdated />
 
-        <h3>More info</h3>
+        <h3>
+          <span aria-hidden="true">*</span>More info
+        </h3>
         <p className="smalltext">
-          This is currently a very rough metric. “Vaccinated Canadians” is different than “vaccines
-          administered”. Check out{' '}
-          <Link href="/methodologies">
-            <a>methodologies</a>
-          </Link>{' '}
-          for more of a discussion.
+          “Vaccinated” is a little ambiguous. Both vaccines currently used in Canada require 2 doses
+          (several weeks apart) to be fully effective. However, receiving 1 dose is partially
+          effective, and indicates how quickly we are dispensing vaccines.
+        </p>
+        <p className="smalltext">
+          The smaller number tracks how many Canadians have received both doses. The larger number
+          tracks how many Canadians have received at least 1 dose. For a more thorough write-up,
+          check out the{' '}
+          <Link href="/methodology">
+            <a>Methodology</a>
+          </Link>
+          .
         </p>
       </section>
 
@@ -118,15 +146,14 @@ const Home = () => (
           <span className="visuallyHidden">Total vaccines administered in Canada </span>By region
         </h2>
         <p>
-          Number of administered vaccines as a percentage of each of Canada’s provinces and
-          territories.
+          Percentage of Canadians who have received vaccines across all provinces and territories.
         </p>
 
         <div className="chart">
           <VictoryChart height={400} width={360} domainPadding={8} theme={theme}>
             <VictoryLegend
               colorScale={[colors.QcOrangeAccent, colors.QcBlueLight]}
-              data={[{ name: 'Vaccinations administered' }, { name: 'Days in 2021' }]}
+              data={[{ name: 'Canadians vaccinated*' }, { name: 'Days in 2021' }]}
             />
             <VictoryAxis
               fixLabelOverlap={true}
@@ -163,18 +190,19 @@ const Home = () => (
 
         <LastUpdated />
 
-        <h3>More info</h3>
+        <h3>
+          <span aria-hidden="true">*</span>More info
+        </h3>
         <p className="smalltext">
-          Some regions have started to administer second doses, while others have not. This graph
-          will change over the next several days as more granular data is released.
+          “Vaccinated” is a little ambiguous. Both vaccines currently used in Canada require 2 doses
+          (several weeks apart) to be fully effective. However, receiving 1 dose is partially
+          effective, and indicates how quickly we are dispensing vaccines.
+        </p>
+        <p className="smalltext">
+          The percentages above track how many Canadians from each region have received at least 1
+          dose.
         </p>
       </section>
-
-      <style jsx>{`
-        .chart + p {
-          margin-top: -30px;
-        }
-      `}</style>
     </div>
   </Layout>
 )